reorganize to fix build errors
